@@ -7,10 +7,15 @@
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
     "bootstrap": "^5.1.3",
-    "react": "^17.0.2",
+    "emailjs-com": "^3.2.0",
+    "react": "^18.3.1",
     "react-bootstrap": "^2.0.3",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.3.1",
+    "react-helmet-async": "^2.0.5",
+    "react-icons": "^5.3.0",
+    "react-router-dom": "^6.27.0",
     "react-scripts": "5.0.0",
+    "typewriter-effect": "^2.21.0",
     "web-vitals": "^2.1.2"
   },
   "scripts": {
