@@ -1,11 +1,9 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from "react-dom/client";
+import App from './app/App';
 import './index.css';
-import App from './App';
 
-ReactDOM.render(
-  <React.StrictMode>
+const root = ReactDOM.createRoot(document.getElementById("root"));
+root.render(
     <App />
-  </React.StrictMode>,
-  document.getElementById('root')
 );
